+++ src/main/java/nextstep/subway/auth/application/UserDetailsService.java
@@ -1,4 +1,4 @@
-package nextstep.subway.auth.ui;
+package nextstep.subway.auth.application;
 
 import nextstep.subway.auth.domain.User;
